Rename `Collisions` to `ContactGraph` and add `Collisions` `SystemParam`
Also moved contact types to `contact_types` module and improved some docs.
